LightSpeech/requirements.txt
torch>=1.2.0
pyworld
pysptk
matplotlib
librosa
tqdm
pandas
numba==0.48.0
numpy
h5py==2.8
requests
scipy==1.3
sympy
PyYAML
praat-parselmouth
tensorboardX
pillow
pyloudnorm
pytorch_lightning==0.6.0
setuptools>=41.0.0
miditoolkit==0.1.7
sklearn
uuid
g2p_en
resemblyzer
webrtcvad
learn2learn
tensorboard==1.15.0
scikit-learn==0.22.0
scikit-image==0.16.2
pypinyin
textgrid
pydub
horology
parallel_wavegan==0.4.0